+++ clicks/digiisolator3/lib_digiisolator3/include/digiisolator3.h
@@ -1,0 +1,212 @@
+/****************************************************************************
+** Copyright (C) 2020 MikroElektronika d.o.o.
+** Contact: https://www.mikroe.com/contact
+**
+** Permission is hereby granted, free of charge, to any person obtaining a copy
+** of this software and associated documentation files (the "Software"), to deal
+** in the Software without restriction, including without limitation the rights
+** to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+** copies of the Software, and to permit persons to whom the Software is
+** furnished to do so, subject to the following conditions:
+** The above copyright notice and this permission notice shall be
+** included in all copies or substantial portions of the Software.
+**
+** THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+** EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
+** OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
+** IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
+** DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT
+** OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE
+**  USE OR OTHER DEALINGS IN THE SOFTWARE.
+****************************************************************************/
+
+/*!
+ * @file digiisolator3.h
+ * @brief This file contains API for DIGI Isolator 3 Click Driver.
+ */
+
+#ifndef DIGIISOLATOR3_H
+#define DIGIISOLATOR3_H
+
+#ifdef __cplusplus
+extern "C"{
+#endif
+
+/**
+ * Any initialization code needed for MCU to function properly.
+ * Do not remove this line or clock might not be set correctly.
+ */
+#ifdef PREINIT_SUPPORTED
+#include "preinit.h"
+#endif
+
+#ifdef MikroCCoreVersion
+    #if MikroCCoreVersion >= 1
+        #include "delays.h"
+    #endif
+#endif
+
+#include "drv_digital_out.h"
+#include "drv_digital_in.h"
+#include "drv_uart.h"
+
+/*!
+ * @addtogroup digiisolator3 DIGI Isolator 3 Click Driver
+ * @brief API for configuring and manipulating DIGI Isolator 3 Click driver.
+ * @{
+ */
+
+/**
+ * @defgroup digiisolator3_cmd DIGI Isolator 3 Device Settings
+ * @brief Settings for DIGI Isolator 3 Click driver.
+ */
+
+/**
+ * @addtogroup digiisolator3_cmd
+ * @{
+ */
+
+/**
+ * @brief DIGI Isolator 3 driver buffer size.
+ * @details Specified size of driver ring buffer.
+ * @note Increase buffer size if needed.
+ */
+#define DIGIISOLATOR3_TX_DRV_BUFFER_SIZE    100
+#define DIGIISOLATOR3_RX_DRV_BUFFER_SIZE    300
+
+/*! @} */ // digiisolator3_cmd
+
+/**
+ * @defgroup digiisolator3_map DIGI Isolator 3 MikroBUS Map
+ * @brief MikroBUS pin mapping of DIGI Isolator 3 Click driver.
+ */
+
+/**
+ * @addtogroup digiisolator3_map
+ * @{
+ */
+
+/**
+ * @brief MikroBUS pin mapping.
+ * @details Mapping pins of DIGI Isolator 3 Click to the selected MikroBUS.
+ */
+#define DIGIISOLATOR3_MAP_MIKROBUS( cfg, mikrobus ) \
+    cfg.tx_pin = MIKROBUS( mikrobus, MIKROBUS_TX ); \
+    cfg.rx_pin = MIKROBUS( mikrobus, MIKROBUS_RX );
+
+/*! @} */ // digiisolator3_map
+/*! @} */ // digiisolator3
+
+/**
+ * @brief DIGI Isolator 3 Click context object.
+ * @details Context object definition of DIGI Isolator 3 Click driver.
+ */
+typedef struct
+{
+    // Modules
+    uart_t uart;                    /**< UART driver object. */
+
+    // Buffers
+    uint8_t uart_rx_buffer[ DIGIISOLATOR3_RX_DRV_BUFFER_SIZE ];  /**< RX Buffer size. */
+    uint8_t uart_tx_buffer[ DIGIISOLATOR3_TX_DRV_BUFFER_SIZE ];  /**< TX Buffer size. */
+
+} digiisolator3_t;
+
+/**
+ * @brief DIGI Isolator 3 Click configuration object.
+ * @details Configuration object definition of DIGI Isolator 3 Click driver.
+ */
+typedef struct
+{
+    // Communication gpio pins
+    pin_name_t rx_pin;              /**< RX pin. */
+    pin_name_t tx_pin;              /**< TX pin. */
+
+    // Static variable
+    uint32_t         baud_rate;     /**< Clock speed. */
+    bool             uart_blocking; /**< Wait for interrupt or not. */
+    uart_data_bits_t data_bit;      /**< Data bits. */
+    uart_parity_t    parity_bit;    /**< Parity bit. */
+    uart_stop_bits_t stop_bit;      /**< Stop bits. */
+
+} digiisolator3_cfg_t;
+
+/**
+ * @brief DIGI Isolator 3 Click return value data.
+ * @details Predefined enum values for driver return values.
+ */
+typedef enum
+{
+    DIGIISOLATOR3_OK = 0,
+    DIGIISOLATOR3_ERROR = -1
+
+} digiisolator3_return_value_t;
+
+/*!
+ * @addtogroup digiisolator3 DIGI Isolator 3 Click Driver
+ * @brief API for configuring and manipulating DIGI Isolator 3 Click driver.
+ * @{
+ */
+
+/**
+ * @brief DIGI Isolator 3 configuration object setup function.
+ * @details This function initializes Click configuration structure to initial
+ * values.
+ * @param[out] cfg : Click configuration structure.
+ * See #digiisolator3_cfg_t object definition for detailed explanation.
+ * @return Nothing.
+ * @note The all used pins will be set to unconnected state.
+ */
+void digiisolator3_cfg_setup ( digiisolator3_cfg_t *cfg );
+
+/**
+ * @brief DIGI Isolator 3 initialization function.
+ * @details This function initializes all necessary pins and peripherals used
+ * for this Click board.
+ * @param[out] ctx : Click context object.
+ * See #digiisolator3_t object definition for detailed explanation.
+ * @param[in] cfg : Click configuration structure.
+ * See #digiisolator3_cfg_t object definition for detailed explanation.
+ * @return @li @c  0 - Success,
+ *         @li @c -1 - Error.
+ * See #err_t definition for detailed explanation.
+ * @note None.
+ */
+err_t digiisolator3_init ( digiisolator3_t *ctx, digiisolator3_cfg_t *cfg );
+
+/**
+ * @brief DIGI Isolator 3 data writing function.
+ * @details This function writes a desired number of data bytes by using UART serial interface.
+ * @param[in] ctx : Click context object.
+ * See #digiisolator3_t object definition for detailed explanation.
+ * @param[in] data_in : Data buffer for sending.
+ * @param[in] len : Number of bytes for sending.
+ * @return @li @c  >=0 - Success,
+ *         @li @c   <0 - Error.
+ * See #err_t definition for detailed explanation.
+ * @note None.
+ */
+err_t digiisolator3_generic_write ( digiisolator3_t *ctx, uint8_t *data_in, uint16_t len );
+
+/**
+ * @brief DIGI Isolator 3 data reading function.
+ * @details This function reads a desired number of data bytes by using UART serial interface.
+ * @param[in] ctx : Click context object.
+ * See #digiisolator3_t object definition for detailed explanation.
+ * @param[out] data_out : Output read data.
+ * @param[in] len : Number of bytes to be read.
+ * @return @li @c  >0 - Number of data bytes read,
+ *         @li @c <=0 - Error/Empty Ring buffer.
+ * See #err_t definition for detailed explanation.
+ * @note None.
+ */
+err_t digiisolator3_generic_read ( digiisolator3_t *ctx, uint8_t *data_out, uint16_t len );
+
+#ifdef __cplusplus
+}
+#endif
+#endif // DIGIISOLATOR3_H
+
+/*! @} */ // digiisolator3
+
+// ------------------------------------------------------------------------ END