+++ clicks/flash13/example/main.c
@@ -1,0 +1,132 @@
+/*!
+ * @file main.c
+ * @brief Flash 13 Click example
+ *
+ * # Description
+ * This example demonstrates the use of Flash 13 Click board by writing specified data to
+ * the memory and reading it back.
+ *
+ * The demo application is composed of two sections :
+ *
+ * ## Application Init
+ * Initializes the driver and checks the communication by reading and verifying the device ID.
+ *
+ * ## Application Task
+ * Writes a desired number of bytes to the memory and then verifies if it is written correctly
+ * by reading from the same memory location and displaying the memory content on the USB UART.
+ * The whole 4KB sector of memory that contains the STARTING_ADDRESS will be erased before writing data.
+ *
+ * @author Stefan Filipovic
+ *
+ */
+
+#include "board.h"
+#include "log.h"
+#include "flash13.h"
+
+#define DEMO_TEXT_MESSAGE_1         "MIKROE"
+#define DEMO_TEXT_MESSAGE_2         "Flash 13 Click"
+#define STARTING_ADDRESS            0x012345ul
+
+static flash13_t flash13;
+static log_t logger;
+
+void application_init ( void )
+{
+    log_cfg_t log_cfg;  /**< Logger config object. */
+    flash13_cfg_t flash13_cfg;  /**< Click config object. */
+
+    /** 
+     * Logger initialization.
+     * Default baud rate: 115200
+     * Default log level: LOG_LEVEL_DEBUG
+     * @note If USB_UART_RX and USB_UART_TX 
+     * are defined as HAL_PIN_NC, you will 
+     * need to define them manually for log to work. 
+     * See @b LOG_MAP_USB_UART macro definition for detailed explanation.
+     */
+    LOG_MAP_USB_UART( log_cfg );
+    log_init( &logger, &log_cfg );
+    log_info( &logger, " Application Init " );
+
+    // Click initialization.
+    flash13_cfg_setup( &flash13_cfg );
+    FLASH13_MAP_MIKROBUS( flash13_cfg, MIKROBUS_1 );
+    if ( SPI_MASTER_ERROR == flash13_init( &flash13, &flash13_cfg ) )
+    {
+        log_error( &logger, " Communication init." );
+        for ( ; ; );
+    }
+    
+    if ( FLASH13_ERROR == flash13_check_communication ( &flash13 ) )
+    {
+        log_error( &logger, " Check communication." );
+        for ( ; ; );
+    }
+
+    log_info( &logger, " Application Task " );
+}
+
+void application_task ( void )
+{
+    uint8_t data_buf[ 128 ] = { 0 };
+
+    log_printf ( &logger, " Memory address: 0x%.6LX\r\n", ( uint32_t ) STARTING_ADDRESS );
+    if ( FLASH13_OK == flash13_erase_memory ( &flash13, FLASH13_CMD_SECTOR_ERASE_4KB, STARTING_ADDRESS ) )
+    {
+        log_printf ( &logger, " Erase memory block (4KB)\r\n" );
+    }
+    memcpy ( data_buf, DEMO_TEXT_MESSAGE_1, strlen ( DEMO_TEXT_MESSAGE_1 ) );
+    if ( FLASH13_OK == flash13_memory_write ( &flash13, STARTING_ADDRESS, data_buf, sizeof ( data_buf ) ) )
+    {
+        log_printf ( &logger, " Write data: %s\r\n", data_buf );
+        Delay_ms ( 100 );
+    }
+    memset ( data_buf, 0, sizeof ( data_buf ) );
+    if ( FLASH13_OK == flash13_memory_read ( &flash13, STARTING_ADDRESS, data_buf, sizeof ( data_buf ) ) )
+    {
+        log_printf ( &logger, " Read data: %s\r\n\n", data_buf );
+        Delay_ms ( 1000 );
+        Delay_ms ( 1000 );
+        Delay_ms ( 1000 );
+    }
+
+    log_printf ( &logger, " Memory address: 0x%.6LX\r\n", ( uint32_t ) STARTING_ADDRESS );
+    if ( FLASH13_OK == flash13_erase_memory ( &flash13, FLASH13_CMD_SECTOR_ERASE_4KB, STARTING_ADDRESS ) )
+    {
+        log_printf ( &logger, " Erase memory block (4KB)\r\n" );
+    }
+    memcpy ( data_buf, DEMO_TEXT_MESSAGE_2, strlen ( DEMO_TEXT_MESSAGE_2 ) );
+    if ( FLASH13_OK == flash13_memory_write ( &flash13, STARTING_ADDRESS, data_buf, sizeof ( data_buf ) ) )
+    {
+        log_printf ( &logger, " Write data: %s\r\n", data_buf );
+        Delay_ms ( 100 );
+    }
+    memset ( data_buf, 0, sizeof ( data_buf ) );
+    if ( FLASH13_OK == flash13_memory_read ( &flash13, STARTING_ADDRESS, data_buf, sizeof ( data_buf ) ) )
+    {
+        log_printf ( &logger, " Read data: %s\r\n\n", data_buf );
+        Delay_ms ( 1000 );
+        Delay_ms ( 1000 );
+        Delay_ms ( 1000 );
+    }
+}
+
+int main ( void ) 
+{
+    /* Do not remove this line or clock might not be set correctly. */
+    #ifdef PREINIT_SUPPORTED
+    preinit();
+    #endif
+    
+    application_init( );
+    
+    for ( ; ; ) 
+    {
+        application_task( );
+    }
+
+    return 0;
+}
+
+// ------------------------------------------------------------------------ END